updates and I see some bitch be touching hats

+++ code/modules/reagents/reagent_containers/glass.dm
@@ -94,8 +94,6 @@
 		if (istype(target, /turf/floor/beach/water))
 			return
 		if (reagents && reagents.total_volume && !istype(src, /obj/item/weapon/reagent_containers/glass/small_pot))
-			playsound(src,'sound/effects/Splash_Small_01_mono.ogg',50,1)
-			user << "<span class='notice'>You splash the solution onto [target].</span>"
 			if (reagents.has_reagent("petroleum", 5))
 				new/obj/effect/decal/cleanable/blood/oil(user.loc)
 			else if (reagents.has_reagent("gasoline", 5))
@@ -106,7 +104,14 @@
 				new/obj/effect/decal/cleanable/blood/oil(user.loc)
 			else if (reagents.has_reagent("olive_oil", 15))
 				new/obj/effect/decal/cleanable/blood/oil(user.loc)
+			else if (reagents.has_reagent("fat_oil", 15))
+				new/obj/effect/decal/cleanable/blood/oil(user.loc)
+
+			var/previous_total_volume = reagents.total_volume
 			reagents.splash(target, reagents.total_volume)
+			if(reagents.total_volume != previous_total_volume) //really splashed
+				playsound(src,'sound/effects/Splash_Small_01_mono.ogg',50,1)
+				user << "<span class='notice'>You splash the solution onto [target].</span>"
 			return
 
 	attackby(obj/item/weapon/W as obj, mob/user as mob)
@@ -157,6 +162,19 @@
 
 			user << "You smash the olives, producing olive oil."
 			reagents.add_reagent("olive_oil", 6)
+			qdel(W)
+			return
+
+		else if (istype(W, /obj/item/weapon/reagent_containers/food/snacks/animalfat))
+
+			if (!is_open_container())
+				user << "<span class='notice'>\The [src] is closed.</span>"
+				return
+			if (!reagents.get_free_space())
+				user << "<span class='notice'>[src] is full.</span>"
+				return
+			user << "You smash the animal fat inside the container, creating lard."
+			reagents.add_reagent("lard", 10)
 			qdel(W)
 			return
 
@@ -711,6 +729,18 @@
 		..()
 		reagents.add_reagent("olive_oil",200)
 
+/obj/item/weapon/reagent_containers/glass/barrel/fat_oil
+	name = "fat oil barrel"
+	desc = "A barrel filled with fat oil."
+	icon = 'icons/obj/barrel.dmi'
+	icon_state = "barrel_wood_drinks"
+	amount_per_transfer_from_this = 10
+	volume = 250
+	density = TRUE
+	New()
+		..()
+		reagents.add_reagent("fat_oil",200)
+
 
 /obj/item/weapon/reagent_containers/glass/barrel/gunpowder
 	name = "gunpowder barrel"
@@ -830,3 +860,20 @@
 			for(var/i=1 to RG.reagents.reagent_list.len)
 				user << "<font color='yellow'><i><b>[RG.reagents.reagent_list[i].name]: </b>[RG.reagents.reagent_list[i].volume] units</i></font>"
 		..()
+
+/obj/item/weapon/reagent_containers/glass/attackby(obj/item/weapon/W as obj, mob/user as mob) //Lard candle making
+	if(istype(W, /obj/item/stack/material/rope))
+		if (reagents.get_reagent_amount("lard") >= 5)
+			var/obj/item/stack/material/rope/R = W
+			visible_message("[user] starts to dip the [R.name] into the [src.name], shaping a candle.")
+			if(do_after(user, 40, user))
+				reagents.remove_reagent("lard", 5)
+				new/obj/item/weapon/flame/candle/lard(user.loc)
+				if(R.amount == 1)
+					qdel(R)
+				else
+					R.amount -= 1
+				return
+	else
+		return ..()
+
